Expand table rows with Stimulus controller
The Expandable stimulus controller accepts some targets on which it
should be able to expand, by initializing their states to collapsed
using the collapsedClass css name.

The `expand()` action simply removes the collapsedClass name from all
the targets.

I considered using toggle here, but that would make it difficult to copy
the text (which is a pretty normal use case I would say for the table
content).