Remove partial index
This seems to be bugging out for sqlite3 adapter using the current
Rails version, by strangely appending “/*application WebhookDelivere */“
to the query.

This could be due to the fact that I passed in a value, whereas the
Rails database migration documentation for partial indexes omits the
value, and just does `where: “active”`. 

https://apidock.com/rails/ActiveRecord/ConnectionAdapters/SchemaStatements/add_index

Nontheless, a bit funky, and note a huge deal to omit for the time being