Move queue solid database schema to primary db
I discovered that rake db:drop will cause issues when re-creating the
databases on a multi database config, causing the queue schema not to
load its migrations and we end up with an empty database.

This is not affected by SolidQueues migrations being faulty, it seems to
be related to the multi database setups only running the migrations for
the primary database. I discovered this while playing around with
SolidCable and decided to backtrack and fix this issue in main branch.

This is due to a bug in Rails: https://github.com/rails/rails/issues/52829
It was fixed by https://github.com/rails/rails/pull/53320 but is part of
Rails main and not 8.0.0.rc2.

Revert this commit when a new release candidate is released including
the https://github.com/rails/rails/commit/ecbdfbc2127f6788c2691d4b02da87c93e0b3b9c
commit.